stub out put vectors sink
